+++ examples/radio-item.php
@@ -23,7 +23,7 @@ $menu = (new CliMenuBuilder)
     })
     ->addSubMenu('Interpreted', function (CliMenuBuilder $b) use ($itemCallable) {
         $b->setTitle('Interpreted Languages')
-            ->radioStyle(function (RadioStyle $style) {
+            ->modifyRadioStyle(function (RadioStyle $style) {
                 $style->setMarkerOff('[ ] ')
                     ->setMarkerOn('[✔] ');
             })
